Add basic measurements history screen
