+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/Koin.kt
@@ -8,6 +8,7 @@ import org.koin.dsl.module
 import org.noiseplanet.noisecapture.permission.defaultPermissionModule
 import org.noiseplanet.noisecapture.permission.platformPermissionModule
 import org.noiseplanet.noisecapture.services.servicesModule
+import org.noiseplanet.noisecapture.ui.features.history.historyModule
 import org.noiseplanet.noisecapture.ui.features.home.homeModule
 import org.noiseplanet.noisecapture.ui.features.measurement.measurementModule
 import org.noiseplanet.noisecapture.ui.features.permission.requestPermissionModule
@@ -35,6 +36,7 @@ fun initKoin(
             requestPermissionModule,
             measurementModule,
             settingsModule,
+            historyModule,
 
             module {
                 includes(additionalModules)
